JIT: x64: ensure writable code memory after alloc
Call JitEnableWrite() immediately following the allocation of the JIT
code memory in the x64 compiler to make certain that it is in RW mode.

This call is made following allocation in the A64 compiler already, see
890844f ("hopefully fix macos").

Switching the memory protection between RW and RX is required on systems
that enforce an executable-space protection policy, such as NetBSD and
OpenBSD, with PaX MPROTECT and W^X respectively.

However, adherence to these policies for these platforms is not
implemented yet, so at this moment in time this call should be a no-op
on x64.